requirements.txt
numpy==1.19.2
tqdm==4.49.0
tensorflow-gpu==1.14.0
keras==2.2.4
Pillow==7.2.0
matplotlib==3.3.2
torch==1.2.0
cupy-cuda100==7.1.1
scikit-image==0.17.2
torchvision==0.4.0
easydict==1.9
tensorboardX==2.0
